Huge code refactoring. I'd recommend skimming the entire project first (start at MainApp.java) to get an idea of what I've done. Basically I split all the test code that's been hacked together over time into their own areas of the application. We can create DataPanels now (for holding the course map and JFreeCharts) that automatically pull in the global data collections necessary to log data. We can also pull in other global information in the future using this approach (such as the soon-to-be JSlider's current position for rendering a unified time from the log in the GUI across all graphs).
Data feeds are lists of sample objects. AccelFeed isn't anything special, but GpsFeed has built-in functionality for computing min/max coordinates for it's contained elements. I rewrote the calculation of these min/max coordinates so that they're computed as GpsSample objects are added to the feed. This prevents the need for bulky linear scans of the feed to determine min/max at render time. All feeds are held by a new global DataFeed singleton object. This way as soon as we need the log data, it's pulled into memory from the file using the LogFileReader and held by the DataFeed singleton for anyone to access.

That's all I can think of for now. Skim all the code for a better idea of what's up.